requirements.txt
discord.py==2.6.4
protobuf==6.33.2
python-dotenv==1.2.1
